Rename env → base.
