+++ src/interpreter/mod.rs
@@ -121,7 +121,7 @@ impl<U> From<U> for Error where U: UserError + Sized {
 }
 
 mod emscripten;
-mod env_native;
+mod native;
 mod imports;
 mod memory;
 mod module;
@@ -143,5 +143,5 @@ pub use self::table::TableInstance;
 pub use self::program::ProgramInstance;
 pub use self::value::RuntimeValue;
 pub use self::variable::{VariableInstance, VariableType, ExternalVariableValue};
-pub use self::env_native::{env_native_module, UserDefinedElements, UserFunctionExecutor, UserFunctionDescriptor};
+pub use self::native::{native_module, UserDefinedElements, UserFunctionExecutor, UserFunctionDescriptor};
 pub use self::emscripten::EnvParams;
